Create Calculate and display total sales
Write a program that will allow a manager to determine and display the total sales for
the department. The manager will enter the sales goal for the month. Then the manager
will enter (input) the sales for each salesperson for four (4) weeks. After each
salesperson, the program should ask the manager if there is another. Once all the sales
are entered the program will accumulate the total sales for the department.
Display the total number of employees in the department and total sales for the
department. The manager earns a 2% bonus on all sales for the month. If the total
department’s sales exceed the sales goal, then the manager’s bonus is 5% of the total
sales. Determine and display the sales goal and manager’s bonus amount.
Hint: two loops are required – use a loop to allow for another salesperson’s monthly
sales to be entered and a second loop for the four (4) weeks of sales of the individual
salesperson.
Some sample Input/Output is provided on the next page. The output is based on the
input for two (2) employees one with weekly sales of (1200, 1300, 1250, 2250) and
second salesperson’s sales (2399, 2103, 1900, 1000). Sales goal for the month -
$11000.
You do not need to round or truncate your output
Save your program as Lab3.py and upload your source file to the appropriate dropbox
in GradeScope, NOT D2L!
Input
Sales goal: 11000
Salesperson 1 week 1: 1200
...
Salesperson 1 week 4: 2250
Another salesperson?: y
Salesperson 2 week 1: 2399
...
Salesperson 2 week 4: 1000
Another salesperson?: n
Output
Department Monthly Sales and Commission
Number of Employees: 2
Department Sales Goal: $11000
Total Sales: $13, 402.00
Mgr. Bonus: $670.10