Create Cost of vacation
Write a program that will allow a manager to determine and the cost of a resort
vacation. A resort client will be able to book a vacation, dining and special excursions.
The client (user) will enter the room selection and the number of days of the vacation.
The room options are:
 Two Queen Beds - $375/night
 One King Bed - $350/night
 Queen Suite - $525/night
 King Suite - $475/night
Meal Options:
 Brunch - $25 per meal
 Dinner - $75 per meal
(15% gratuity will automatically be added to the charges)
Excursions: (multiple excursions can be selected)
 Picnic - $50
 Snorkeling - $25 per person
 Guided Hike - $17 per person
 Boat Dinner - $200
When writing this program please include (at least) the following functions/methods.
A roomCost function which will accept the number of nights and the room type
(these values will be passed to the function) and the total cost will be returned to
the calling program
A mealCost function this function will receive the number of brunches and
dinners through the parameter list. The function will return the total meal cost
(including gratuity) to the calling program.
An excursionCost function which will accept the number of each event
requested and return the total cost of excursions to the calling program.
The program will display the number of nights of the stay and the room cost, the meal
cost and excursion cost. A total for all items selected will be displayed. You do not need
to truncate or round your output.
Save your program as Lab4.py and upload your source file to the appropriate dropbox
in GradeScope, NOT D2L!
Output Example
Input
Number of nights: 5
Number of people: 2
Room Types:
(1) - Two Queen Beds
(2) - One King Bed
(3) - Queen Suite
(4) - King Suit
Select Type: 1
How many Brunches: 5
How many Dinners: 5
Picnic?: y
Snorkeling?: y
Guided Hike?: y
Boat Dinner?: y
Output
Room Cost: $1875
Meal Cost: $575
Excursion Cost: $334
Total Cost: $278